+++ packages/request-utils/package.json
@@ -16,10 +16,12 @@
 	"license": "ISC",
 	"dependencies": {
 		"express": "^4.18.2",
+		"formidable": "^3.5.2",
 		"zod": "^3.24.1"
 	},
 	"devDependencies": {
 		"@types/express": "^5.0.0",
-		"@types/express-serve-static-core": "^5.0.6"
+		"@types/express-serve-static-core": "^5.0.6",
+		"@types/formidable": "^3.4.5"
 	}
 }
